requirements.txt
asttokens==2.4.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
Flask==3.0.3
greenlet==3.0.3
idna==3.7
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.0
jupyter_core==5.6.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mysql-connector-python==8.3.0
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pyzmq==25.1.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.12
Werkzeug==3.0.3
